refactor: threadとannouncementでコード分割
